use nameValidationFormat() everywhere
